`refactor(command-service): extract Spotify and playlist logic into dedicated services`
